requirements.txt
Django==1.11.17
Pillow==6.1.0
pytz==2019.1
django-paypal==1.0